Remove unnecessary abstractions
